+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Development
 
+### Changed
+- Calculations for available screen width and height to now include screen density and be more independent of physical pixel
+
 ## v1.18.0
 
 ### Added